refactor(ui): split placeholder store into types/constants/utils files
Match the file-per-concern layout the rest of the feature follows
(useVectorEmbeddingCollapse.{types,constants,utils}). Move the interfaces
to vectorEmbeddingPlaceholders.types.ts (barrelled like the sibling types
file), the regex to .constants.ts, and the CSPRNG session-id helper to
.utils.ts; the store file now holds only state and the public methods.
