+++ redisinsight/ui/src/utils/monaco/index.ts
@@ -2,6 +2,7 @@ export * from './monacoUtils'
 export * from './vectorEmbeddingUtils'
 export * from './vectorEmbeddingUtils.types'
 export * from './vectorEmbeddingPlaceholders'
+export * from './vectorEmbeddingPlaceholders.types'
 export * from './monacoInterfaces'
 export * from './monacoRedisCompletionProvider'
 export * from './monacoRedisMonarchTokensProvider'
